requirements.txt
flask
requests
beautifulsoup4
openai
pinecone-client
flask-cors
python-dotenv
chromadb
langchain-huggingface
langchain-community langchain-core